backend/requirements.txt
asgiref==3.7.2
Django==5.0.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
gunicorn==21.2.0
packaging==23.2
pillow==10.2.0
PyJWT==2.8.0
pytz==2023.4
rest-framework-simplejwt==0.0.2
sqlparse==0.4.4
typing_extensions==4.9.0
